# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Orbital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Satellite {
  id: number;
  noradId: number;
  name: string;
  orbitType: string;
  /** @nullable */
  altitude: number | null;
  /** @nullable */
  inclination: number | null;
  /** @nullable */
  velocity?: number | null;
  epoch: string;
  /** @nullable */
  eccentricity?: number | null;
  /** @nullable */
  raan?: number | null;
  lastUpdated: string;
}

export interface SatelliteDetail {
  id: number;
  noradId: number;
  name: string;
  orbitType: string;
  /** @nullable */
  altitude: number | null;
  /** @nullable */
  inclination: number | null;
  /** @nullable */
  velocity?: number | null;
  epoch: string;
  /** @nullable */
  eccentricity?: number | null;
  /** @nullable */
  raan?: number | null;
  lastUpdated: string;
  /** @nullable */
  riskScore: number | null;
  /** @nullable */
  riskCategory?: string | null;
}

export interface SatelliteListResponse {
  data: Satellite[];
  total: number;
  page: number;
  limit: number;
}

export interface DebrisObject {
  id: number;
  noradId: number;
  name: string;
  objectType: string;
  /** @nullable */
  altitude: number | null;
  /** @nullable */
  inclination: number | null;
  epoch: string;
  /** @nullable */
  eccentricity?: number | null;
  lastUpdated: string;
}

export interface DebrisListResponse {
  data: DebrisObject[];
  total: number;
  page: number;
  limit: number;
}

export interface OrbitBand {
  orbitType: string;
  count: number;
  percentage: number;
}

export interface AnalyticsOverview {
  totalSatellites: number;
  totalDebris: number;
  totalRocketBodies: number;
  highRiskObjects: number;
  criticalRiskObjects?: number;
  /** @nullable */
  lastRefresh: string | null;
  orbitBreakdown: OrbitBand[];
  recentCollisionWarnings?: number;
}

export interface AltitudeBand {
  altitudeKm: number;
  count: number;
  label: string;
}

export interface InclinationBand {
  inclinationDeg: number;
  count: number;
}

export interface CongestionHistoryPoint {
  date: string;
  totalObjects: number;
  satellites: number;
  debris: number;
}

export interface CongestionData {
  altitudeBands: AltitudeBand[];
  densityByInclination: InclinationBand[];
  history: CongestionHistoryPoint[];
}

export interface ForecastPoint {
  date: string;
  predicted: number;
  lower: number;
  upper: number;
}

export interface ForecastData {
  horizon: string;
  dataPoints: ForecastPoint[];
  trend: string;
  confidence: number;
}

export interface RiskScore {
  noradId: number;
  name: string;
  score: number;
  category: string;
  lastCalculated: string;
}

export interface RiskFactors {
  nearbyDebrisCount: number;
  orbitCongestion: number;
  collisionProbability: number;
  satelliteAge: number;
  closeApproachFrequency: number;
}

export interface RiskScoreDetail {
  noradId: number;
  name: string;
  score: number;
  category: string;
  lastCalculated: string;
  factors: RiskFactors;
}

export interface CollisionEvent {
  id: number;
  object1Name: string;
  object2Name: string;
  object1NoradId: number;
  object2NoradId: number;
  missDistanceKm: number;
  riskLevel: string;
  /** @nullable */
  predictedTime: string | null;
  /** @nullable */
  probability?: number | null;
}

export interface CollisionListResponse {
  data: CollisionEvent[];
  total: number;
  highRiskCount: number;
  criticalCount: number;
}

export interface LastUpdateInfo {
  /** @nullable */
  lastRefresh: string | null;
  satelliteCount: number;
  debrisCount: number;
  rocketBodyCount: number;
  /** @nullable */
  nextRefresh?: string | null;
}

export interface RefreshResult {
  success: boolean;
  message: string;
  startedAt: string;
}

export type SystemStatusDataSourceStatus = {[key: string]: string};

export interface SystemStatus {
  schedulerRunning: boolean;
  /** @nullable */
  lastRefresh: string | null;
  /** @nullable */
  nextRefresh: string | null;
  dataSourceStatus: SystemStatusDataSourceStatus;
  apiHealth: string;
}

export type ListSatellitesParams = {
search?: string;
orbitType?: string;
sortBy?: ListSatellitesSortBy;
sortOrder?: ListSatellitesSortOrder;
page?: number;
limit?: number;
};

export type ListSatellitesSortBy = typeof ListSatellitesSortBy[keyof typeof ListSatellitesSortBy];


export const ListSatellitesSortBy = {
  name: 'name',
  altitude: 'altitude',
  inclination: 'inclination',
  noradId: 'noradId',
} as const;

export type ListSatellitesSortOrder = typeof ListSatellitesSortOrder[keyof typeof ListSatellitesSortOrder];


export const ListSatellitesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListDebrisParams = {
type?: ListDebrisType;
search?: string;
page?: number;
limit?: number;
};

export type ListDebrisType = typeof ListDebrisType[keyof typeof ListDebrisType];


export const ListDebrisType = {
  debris: 'debris',
  rocket_body: 'rocket_body',
  all: 'all',
} as const;

export type GetOrbitalForecastParams = {
horizon?: GetOrbitalForecastHorizon;
};

export type GetOrbitalForecastHorizon = typeof GetOrbitalForecastHorizon[keyof typeof GetOrbitalForecastHorizon];


export const GetOrbitalForecastHorizon = {
  '30d': '30d',
  '90d': '90d',
  '1y': '1y',
} as const;

export type ListRiskScoresParams = {
category?: ListRiskScoresCategory;
limit?: number;
};

export type ListRiskScoresCategory = typeof ListRiskScoresCategory[keyof typeof ListRiskScoresCategory];


export const ListRiskScoresCategory = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type ListCollisionsParams = {
riskLevel?: ListCollisionsRiskLevel;
limit?: number;
};

export type ListCollisionsRiskLevel = typeof ListCollisionsRiskLevel[keyof typeof ListCollisionsRiskLevel];


export const ListCollisionsRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;