# lib/api-zod/src/generated/types/collisionEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Orbital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CollisionEvent {
  id: number;
  object1Name: string;
  object2Name: string;
  object1NoradId: number;
  object2NoradId: number;
  missDistanceKm: number;
  riskLevel: string;
  /** @nullable */
  predictedTime: string | null;
  /** @nullable */
  probability?: number | null;
}