# lib/api-zod/src/generated/types/debrisListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Orbital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DebrisObject } from './debrisObject';

export interface DebrisListResponse {
  data: DebrisObject[];
  total: number;
  page: number;
  limit: number;
}