# lib/api-zod/src/generated/types/riskScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Orbital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskScore {
  noradId: number;
  name: string;
  score: number;
  category: string;
  lastCalculated: string;
}