# lib/api-zod/src/generated/types/satelliteDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Orbital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SatelliteDetail {
  id: number;
  noradId: number;
  name: string;
  orbitType: string;
  /** @nullable */
  altitude: number | null;
  /** @nullable */
  inclination: number | null;
  /** @nullable */
  velocity?: number | null;
  epoch: string;
  /** @nullable */
  eccentricity?: number | null;
  /** @nullable */
  raan?: number | null;
  lastUpdated: string;
  /** @nullable */
  riskScore: number | null;
  /** @nullable */
  riskCategory?: string | null;
}