# lib/api-zod/src/generated/types/satelliteListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Orbital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Satellite } from './satellite';

export interface SatelliteListResponse {
  data: Satellite[];
  total: number;
  page: number;
  limit: number;
}